refactor: separate event emitter from its react HOC, add test
