+++ package.json
@@ -22,7 +22,8 @@
     "tx": "tx pull --all || true",
     "watch": "NODE_ENV=development webpack --config ./config/webpack.target.browser.js --watch --display-chunks",
     "watch:mobile": "NODE_ENV=development webpack --config ./config/webpack.target.mobile.js --watch --display-chunks",
-    "shasum": "shasum -a 256 dist/cozy-bar.min.js && shasum -a 256 dist/cozy-bar.min.css"
+    "shasum": "shasum -a 256 dist/cozy-bar.min.js && shasum -a 256 dist/cozy-bar.min.css",
+    "test": "jest"
   },
   "devDependencies": {
     "autoprefixer": "^7.1.4",
@@ -43,6 +44,7 @@
     "deep-equal": "^1.0.1",
     "extract-text-webpack-plugin": "^3.0.1",
     "file-loader": "^1.1.5",
+    "jest": "^21.2.1",
     "json-loader": "^0.5.7",
     "npm-run-all": "^4.1.1",
     "postcss": "^6.0.13",
